Shifted to Gradle 7.0 due to Apple M1: Standard API call and Room wor…

+++ app/src/main/java/com/yachint/twitterdrift/data/model/trends/Trend.kt
@@ -9,5 +9,5 @@ data class Trend(
     @PrimaryKey(autoGenerate = true) var id: Int = 0,
     val name: String,
     val url: String,
-    val tweet_volume: Int
+    val tweet_volume: Long
 ): BaseModel
